feat: Create updateAirTableRow
Refactored recallNeedsRequest to extrapolate and used a more generic
update function. Updated tests. This commit will close #24
